LPS-74461 Renamed JS files
