+++ modules/apps/web-experience/modern-site-building/modern-site-building-page-web/build.gradle
@@ -1,3 +1,7 @@
+configJSModules {
+	include "**/*.js*"
+}
+
 dependencies {
 	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	provided group: "com.liferay", name: "com.liferay.application.list.api", version: "2.0.0"
@@ -15,4 +19,8 @@ dependencies {
 
 node {
 	nodeVersion = "8.4.0"
+}
+
+transpileJS {
+	srcIncludes = "**/*.js"
 }
