+++ modules/apps/web-experience/modern-site-building/modern-site-building-page-web/src/main/resources/META-INF/resources/js/MSBPageList.js
@@ -1,7 +1,7 @@
 import Component from 'metal-component';
 import {Config} from 'metal-state';
 import Soy from 'metal-soy';
-import './MSBPageListBlock.es';
+import './MSBPageListBlock';
 import templates from './MSBPageList.soy';
 
 /**
