+++ modules/apps/web-experience/modern-site-building/modern-site-building-page-web/src/main/resources/META-INF/resources/js/MSBPageListBlock.js
